+++ HomeAccessPlus/HAPi.swift
@@ -454,31 +454,43 @@ class HAPi {
     ///
     /// - author: Jonathan Hart (stuajnht) <stuajnht@users.noreply.github.com>
     /// - since: 0.5.0-alpha
-    /// - version: 1
-    /// - date: 2016-01-05
+    /// - version: 2
+    /// - date: 2016-01-09
     ///
     /// - parameter deviceFileLocation: The path to the file on the device (normally
     ///                                 stored in a folder called "inbox" which can
     ///                                 be found in the documents directory)
     /// - parameter serverFileLocation: The location on the HAP+ server that the file
     ///                                 is going to be uploaded to
-    func uploadFile(deviceFileLocation: NSURL, serverFileLocation: String, callback:(result: Bool, uploading: Bool, uploadedBytes: Int64, totalBytes: Int64) -> Void) -> Void {
+    /// - parameter fileFromPhotoLibrary: Is the file being uploaded coming from the photo
+    ///                                   library on the device, or from another app
+    func uploadFile(deviceFileLocation: NSURL, serverFileLocation: String, fileFromPhotoLibrary: Bool, callback:(result: Bool, uploading: Bool, uploadedBytes: Int64, totalBytes: Int64) -> Void) -> Void {
         // Checking that we still have a connection to the Internet
         if (checkConnection()) {
             // Getting the name of the file that is being uploaded from the
-            // location of the file on the device, which is needed when settings
+            // location of the file on the device, which is needed when setting
             // the httpHeaders
             logger.debug("Location of file on device: \(deviceFileLocation)")
+            logger.debug("File coming from photos library: \(fileFromPhotoLibrary)")
+            
+            var fileName = ""
+            
+            // As the file is coming from an external app, it will be saved
+            // on the device in a physical location with an extension. We
+            // can just split the path and get the file name from the last
+            // array value
             let pathArray = String(deviceFileLocation).componentsSeparatedByString("/")
             
             // Forcing an unwrap of the value, otherwise the file name
             // is Optional("<nale>") which causes the HAP+ server to
             // do a 500 HTTP error
             // See: http://stackoverflow.com/a/25848016
-            var fileName = pathArray.last!
+            fileName = pathArray.last!
+            
             // Removing any encoded characters from the file name, so
             // HAP+ saves the file with the correct file name
             fileName = fileName.stringByRemovingPercentEncoding!
+            
             logger.debug("Name of file being uploaded: \(fileName)")
             
             // Setting the tokens that are collected from the login, so the HAP+